LayoutTests/crypto/webkitSubtle/rsa-postMessage-expected.txt

Test sending crypto keys via postMessage.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


Key received in a message:
PASS key.type is 'private'
PASS key.extractable is true
PASS key.algorithm.name is 'RSASSA-PKCS1-v1_5'
PASS key.algorithm.modulusLength is 2048
PASS bytesToHexString(key.algorithm.publicExponent) is '010001'
PASS key.usages is ["sign", "verify"]

Exported back to JWK:
PASS exportedJWK.kty is 'RSA'
PASS exportedJWK.alg is 'RS256'
PASS exportedJWK.ext is true
PASS exportedJWK.key_ops is ['sign','verify']
PASS exportedJWK.use is undefined
PASS exportedJWK.n is privateKeyJSON.n
PASS exportedJWK.e is privateKeyJSON.e
PASS exportedJWK.d is privateKeyJSON.d
PASS exportedJWK.p is privateKeyJSON.p
PASS exportedJWK.q is privateKeyJSON.q
PASS exportedJWK.dp is privateKeyJSON.dp
PASS exportedJWK.dq is privateKeyJSON.dq
PASS exportedJWK.qi is privateKeyJSON.qi

PASS successfullyParsed is true

TEST COMPLETE
